AS-IS ) 회원가입 사진전, 폴더 만들기 TO-BE ) 사진 데이터 기능 추가
