+++ src/main/java/com/spring/sharepod/dto/request/Board/BoardAllRequestDto.java
@@ -1,4 +1,4 @@
-package com.spring.sharepod.dto.request;
+package com.spring.sharepod.dto.request.Board;
 
 
 import lombok.AllArgsConstructor;
@@ -8,10 +8,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class BoardFilterAndCategoryRequestDto {
-    private String filtertype;
+public class BoardAllRequestDto {
+
+    private String filtering;
     private String category;
     private String mapdata;
+
 
 
 }
