+++ src/main/java/com/spring/sharepod/dto/request/Board/BoardFilterAndCategoryRequestDto.java
@@ -1,4 +1,4 @@
-package com.spring.sharepod.dto.request;
+package com.spring.sharepod.dto.request.Board;
 
 
 import lombok.AllArgsConstructor;
@@ -8,12 +8,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class BoardAllRequestDto {
-
-    private String filtering;
+public class BoardFilterAndCategoryRequestDto {
+    private String filtertype;
     private String category;
     private String mapdata;
-
 
 
 }
