+++ src/main/java/com/spring/sharepod/model/AllVideo.java
@@ -1,6 +1,6 @@
 package com.spring.sharepod.model;
 
-import com.spring.sharepod.dto.response.VideoAllResponseDto;
+import com.spring.sharepod.dto.response.Board.VideoAllResponseDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
