+++ src/main/java/com/spring/sharepod/service/UserService.java
@@ -1,6 +1,6 @@
 package com.spring.sharepod.service;
 
-import com.spring.sharepod.dto.request.UserRegisterRequestDto;
+import com.spring.sharepod.dto.request.User.UserRegisterRequestDto;
 import com.spring.sharepod.entity.User;
 import com.spring.sharepod.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -8,8 +8,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Collections;
 
 
 @Service
